digital_event_organizer/requirements.txt
Flask==2.3.0
Flask-MySQLdb==1.0.1
Flask-Mail==0.9.1
python-dotenv==1.0.0
razorpay==1.3.0
requests==2.31.0
Werkzeug==2.3.0
mysql-connector-python==8.0.33
